Addition deltadownload fanout
Addition of 2 new actions for the project to complete
deletadownlaod will write the delta of two snapshots to a specified raw device location
fanout will distribute a raw device as a snapshot to a .txt list of specified regions

This is a rebase of 4 commits:

Added windows branch (deleted) functionality

Changed the open file read permissions from read and create to write
only. This is the only way to interact with raw devices on Windows and
it is assumed that using minimal permissions will not break unix

Modified download tests to work with windows

Instead of creating a new raw device use the existing small test drive

Fanout Complete. BUG: deltadownload

The fanout CLI action appears to work as intended, a snapshot can be
generated in many regions. TODO: write a canary test to prove that this
functionality works and ensure backwards compatibility with future
modifications
deltadownload currently has bug where when an attribute
'SecondChangedBlock' is not present the causes thread download action to
fail

Deltadownload functionality now works

Needed to modify the get_changed_block() logic to use try_get_block with
second snapshot and SecondBlockToken when possible. Otherwise use first
snapshot and FirstBlockToken