Flexible Snapshot Proxy is Open Sourced
Approval from Amazon Open Source department has been given and necessary
license has been attached - Apache 2.0. As well as additional helpful
supporting documents for contributing and other compliance notices

From this point forward the project is only referred to as
flexible-snapshot-proxy or fsp. An old name for this project was
ebs-playground or ebs

There exists a need to complete the subbed out README for this project
as well as potentially add and complete additional markdown documentation.

Added clarification to Reporting guidelines, fixed minor formatting.

[BUG] currently the tests are broken

Created Change log

Removed unnecessary file regions.txt

Was used in full stack tests. Will write a temp file in setup() as
alternative

New and improved requirements

Fixed dependency checker

Now it tries to install all the packages quietly when script starts and
aborts if this is not possible. Also fixed the singleton setup step to
only use boto3 class and not aws cli calls.

Using old numpy no python3.8

Amazon Linux 2 does not support python3.8 get. A more stable package of
numpy exists as 1.23.0. However it requires python 3.8 to run. As soon
as this blocker is dropped adjust the python requirements file.
REVISE: Not a big issue, the vulnerability does not apply to this code

Added a quick comment on requirements