Refactored Test/
For security: replaced unnecessary os.system() calls with subprocess
library calls for safety

Split tests into functional a.k.a. canary tests and unit tests which are
currently lacking but should be added to as features get developed

Are create implemented a setup script to generate the config.yaml file
needed to run the testing script. todo: still need to complete the
option to have this setup script create all on ec2 resources required
(because the script is being run on ec2)

This is a refactor of 6 commits:

Refactor of test/ folder

Now uses broken into unit and functional tests with a main testing
script file to run any and all tests that are desired. Currently tests
are just stubbed out but will be filled in shortly

Added a yaml requirement to project for test/

The testing configuration file will be in yaml to increase user
readability. Additionally in future when the fsp script is tunable a
yaml config file should be used there too

Wrote unit test for dependency checker

These mainly server as a template for how to write unit tests and add
them into the argument parser found in test.py

Added a testing setup method

todo: iterate on this approach to have aws resources configured and the
config file generated accordingly.

Updated README in testing for refactor