Dependency check cache and suppress_writes
This is a rebase of 2 commits:

Completed Dependency Check Cache

A hidden file is created when the dependency check passes then on next
re-entries to the script, the creation time stamp will be checked to see if
it is older than 7 days and if so run the dependency checker

Created suppress_writes cli arg option

Intended Use Case: Small devices such as an AWS snowcone that do not
have the space to store log files. Log files will be implemented in the
future
Refactored some arg parsing logic to make more readable. And added the
suppress_writes as a project scoped variable in singleton.py